fix(tests): centralize convex-test scheduler-noise filter in setup (#525)
Many Convex mutations/actions schedule side-effect work via
`ctx.scheduler.runAfter(0, …)` — Supabase dual-writes, activity-envelope
fan-out, audit logging, SMS reminders, automation triggers, etc.
convex-test fires those callbacks via `setTimeout(0, …)` against a
captured `global.Convex` reference. Once the next test creates a fresh
instance, orphan callbacks from the previous test land on a stale `db`
whose transaction has already closed and surface as:
  - `Write outside of transaction N;_scheduled_functions`, or
  - `Cannot read properties of null (reading 'state')` (job lookup
    against the new instance's `_scheduled_functions` table returns
    null).

Both are pure test-harness artefacts: every assertion in the originating
test has already passed by the time they fire, but they flip vitest's
exit code via `unhandledRejection`.

Eight test files (#506, #509, #511, #513, #514, #516, #517) had each
grown an identical per-file boilerplate — a `SCHEDULER_NOISE` regex
array, `beforeAll`/`afterAll` to attach/detach a `process.on
("unhandledRejection")` listener, plus an `afterEach` yield to drain
the queue. The per-file approach has a real gap: rejections that fire
*between* files or after a file's `afterAll` has already torn its
listener down still escape and fail the suite (visible in
`appointmentSms.test.ts`, which never had a filter at all because it
relies on `finishAllScheduledFunctions` instead, yet still leaked
unhandled rejections from orphan scheduled jobs).

Move the filter to `tests/convex/_setup.ts` (already loaded via
vitest `setupFiles`) so a single process-wide listener stays attached
for the whole run. Both regex shapes (`Write outside of transaction
N;_scheduled_functions` and `Cannot read properties of null (reading
'state')`) are matched. Unknown rejections still fall through to
vitest's own listener and surface as failures.

The per-file `afterEach` setTimeout(0) yields stay — they're real
queue-drain cleanup, not noise filtering, and reduce the count of
orphan callbacks that have to be swallowed by the global listener.

Confirmed: running the full convex test suite no longer reports the
`Errors  N error` summary line. The remaining test failures in
`appointmentSms.test.ts` / `automation.test.ts` are pre-existing test
rot unrelated to scheduler noise.
