+++ tests/convex/appointmentStateMachine.test.ts
@@ -1,43 +1,14 @@
-import { afterAll, afterEach, beforeAll, beforeEach, describe, expect, test, vi } from "vitest";
+import { afterEach, beforeEach, describe, expect, test, vi } from "vitest";
 import { api } from "../../convex/_generated/api";
 import { createTestCtx, seedTestUser, seedGabinetPrereqs } from "../../convex/_test_helpers";
 import { createSupabaseDb } from "../../convex/_helpers/supabaseDb";
 
-// State-machine tests assert on appointment status only. The action handler
-// writes the status synchronously, but it also `runAfter(0, …)` schedules
-// side-effect jobs (SMS, automation events, reminder cancellation, …).
-// convex-test runs those via `setTimeout(0, …)` against a `global.Convex`
-// reference; once the next test creates a fresh test instance, the orphan
-// callbacks from the previous test fire against the new `global.Convex` and
-// surface as "Write outside of transaction" / null-state unhandled rejections
-// that fail the suite (even though every assertion in the test passed).
-//
-// `finishAllScheduledFunctions` is not safe here either: under the in-memory
-// Supabase mock some side-effect mutations throw (the gabinet completion path
-// inserts into `loyaltyTransactions` with a non-Convex id, which the
-// validator rejects), and the poller crashes on that null state.
-//
-// So we install a per-file unhandledRejection filter that silently consumes
-// ONLY the two known convex-test scheduler error shapes. Unknown rejections
-// fall through and are still surfaced by vitest's own listener.
-const SCHEDULER_NOISE = [
-  /Write outside of transaction \d+;_scheduled_functions/,
-  /Cannot read properties of null \(reading 'state'\)/,
-];
-
-function onUnhandledRejection(reason: unknown) {
-  const msg = reason instanceof Error ? reason.message : String(reason);
-  if (SCHEDULER_NOISE.some((re) => re.test(msg))) return;
-  // Other rejections are left to vitest's own listener.
-}
-
-beforeAll(() => {
-  process.on("unhandledRejection", onUnhandledRejection);
-});
-
-afterAll(() => {
-  process.off("unhandledRejection", onUnhandledRejection);
-});
+// `finishAllScheduledFunctions` is not safe under the in-memory Supabase mock:
+// some side-effect mutations throw (the gabinet completion path inserts into
+// `loyaltyTransactions` with a non-Convex id which the validator rejects), and
+// the poller then crashes on a null `_scheduled_functions` row. Rely instead
+// on the process-wide scheduler-noise filter installed in
+// tests/convex/_setup.ts plus the setTimeout(0) yield below to drain the queue.
 
 function createManagedTestCtx() {
   return createTestCtx();
@@ -56,7 +27,6 @@ beforeEach(() => {
 afterEach(async () => {
   // Let any pending setTimeout(0) side-effect callbacks from the test fire
   // against the *current* instance before the next test creates a new one.
-  // Anything still queued past this yield is handled by the swallower above.
   await new Promise((resolve) => setTimeout(resolve, 0));
   vi.unstubAllGlobals();
 });
