+++ tests/convex/emailActivities.test.ts
@@ -1,4 +1,4 @@
-import { afterAll, afterEach, beforeAll, describe, expect, test, vi } from "vitest";
+import { afterEach, describe, expect, test, vi } from "vitest";
 import { api, internal } from "../../convex/_generated/api";
 import { createTestCtx, seedTestUser } from "../../convex/_test_helpers";
 import { createSupabaseDb } from "../../convex/_helpers/supabaseDb";
@@ -8,39 +8,12 @@ vi.mock("@cvx/email", () => ({
   sendEmail: vi.fn(async () => undefined),
 }));
 
-// `emails.send` and `emails_internal.insertInbound` schedule Supabase
-// dual-writes / activity-envelope side effects via `ctx.scheduler.runAfter(0, …)`
-// (and `emails.send` itself is now an action that runs an internal mutation).
-// convex-test fires those callbacks via `setTimeout(0, …)` against a
-// `global.Convex` reference; once the next test creates a fresh instance,
-// orphan callbacks from the previous test fire against the new `global.Convex`
-// and surface as "Write outside of transaction" unhandled rejections that flip
-// vitest's exit code even though every assertion passes. Match the per-file
-// filter used by payments/packageActivities (#511, #514) — swallow only the
-// known scheduler-noise shape, let everything else through.
-const SCHEDULER_NOISE = [
-  /Write outside of transaction \d+;_scheduled_functions/,
-];
-
-function onUnhandledRejection(reason: unknown) {
-  const msg = reason instanceof Error ? reason.message : String(reason);
-  if (SCHEDULER_NOISE.some((re) => re.test(msg))) return;
-}
-
-beforeAll(() => {
-  process.on("unhandledRejection", onUnhandledRejection);
-});
-
-afterAll(() => {
-  process.off("unhandledRejection", onUnhandledRejection);
-});
-
 afterEach(async () => {
-  // Let any pending setTimeout(0) side-effect callbacks from the test fire
-  // against the *current* instance before the next test creates a new one.
   // `insertInbound` schedules an internalAction (writeEmailToSupabase) whose
   // setTimeout-driven lifecycle (mark in-progress → run → mark success) spans
   // several microtask ticks, so yield repeatedly until the queue settles.
+  // The process-wide scheduler-noise filter in tests/convex/_setup.ts swallows
+  // any orphan-write rejections that still escape.
   for (let i = 0; i < 10; i++) {
     await new Promise((resolve) => setTimeout(resolve, 0));
   }
