+++ tests/convex/payments.test.ts
@@ -1,36 +1,12 @@
-import { afterAll, afterEach, beforeAll, describe, expect, test } from "vitest";
+import { afterEach, describe, expect, test } from "vitest";
 import { api } from "../../convex/_generated/api";
 import { createTestCtx, seedTestUser } from "../../convex/_test_helpers";
-
-// payments.{create,markPaid,refund} schedule audit-log / activity-envelope
-// dual-writes via `ctx.scheduler.runAfter(0, …)`. convex-test fires those via
-// `setTimeout(0, …)` against a `global.Convex` reference; once the next test
-// creates a fresh instance, the orphan callbacks from the previous test fire
-// against the new `global.Convex` and surface as "Write outside of
-// transaction" unhandled rejections that fail the suite even though every
-// assertion passes. Match the per-file filter used by appointmentStateMachine
-// (see #506) — swallow only the known scheduler noise shape, let everything
-// else through to vitest's listener.
-const SCHEDULER_NOISE = [
-  /Write outside of transaction \d+;_scheduled_functions/,
-];
-
-function onUnhandledRejection(reason: unknown) {
-  const msg = reason instanceof Error ? reason.message : String(reason);
-  if (SCHEDULER_NOISE.some((re) => re.test(msg))) return;
-}
-
-beforeAll(() => {
-  process.on("unhandledRejection", onUnhandledRejection);
-});
-
-afterAll(() => {
-  process.off("unhandledRejection", onUnhandledRejection);
-});
 
 afterEach(async () => {
   // Let any pending setTimeout(0) side-effect callbacks from the test fire
   // against the *current* instance before the next test creates a new one.
+  // The process-wide scheduler-noise filter installed in tests/convex/_setup.ts
+  // silently consumes the orphan-write rejections that still escape this yield.
   await new Promise((resolve) => setTimeout(resolve, 0));
 });
